Move name-order routing into pipeline module
