+++ sinonym/pipeline/__init__.py
@@ -1,0 +1,1 @@
+"""Pipeline-oriented helpers that are not part of the detector API."""
